requirements.txt
opencv-python==4.9.0.80
pillow==10.2.0
pyautogui==0.9.54
numpy==1.26.4